refactor(web): decompose chat shell and session helpers
