+++ packages/franken-web/src/components/chat-shell/network-error.ts
@@ -1,0 +1,3 @@
+export function networkErrorMessage(error: unknown, fallback: string): string {
+  return error instanceof Error ? error.message : fallback;
+}
